PathFile

+++ public/js/homeAdminPage.js
@@ -11,6 +11,7 @@ const TheAuthordiv=document.getElementById("TheAuthordiv");
 const addingNewAuthorBtn=document.getElementById("addingNewAuthorBtn");
 const addnewAuthor=document.getElementById("addnewAuthor");
 const AuthorTable=document.getElementById("AuthorTable")
+const inputAuthorBioGraphyBox=document.getElementById("inputAuthorBioGraphyBox")
 const pleaseEnterDataAuthor=document.getElementById("pleaseEnterDataAuthor")
 const TheBookdiv=document.getElementById("TheBookdiv");
 const addingNewBookBtn=document.getElementById("addingNewBookBtn")
@@ -19,14 +20,42 @@ const inputCatTypForTheBookBox=document.getElementById("inputCatTypForTheBookBox
 const addnewbookBtnInTable=document.getElementById("addnewbookBtnInTable")
 const inputAuthorTypForTheBookBox=document.getElementById("inputAuthorTypForTheBookBox")
 const inputBookBoxName=document.getElementById("inputBookBoxName")
+const inputBookBoxDescription=document.getElementById("inputBookBoxDescription")
 const inputImageForTheBookBox=document.getElementById("inputImageForTheBookBox")
 const pleaseEnterDataBook=document.getElementById("pleaseEnterDataBook")
 const bookTable=document.getElementById("bookTable");
+const modalCannotBeDeleted=document.getElementById("modalCannotBeDeleted")
+const uploadBtnBookImage=document.getElementById("uploadBtnBookImage")
+const inputImageForTheAuthorBox=document.getElementById("inputImageForTheAuthorBox")
+
 let theAuthors;    //the  array authores from database 
 let theCartygroies ; //the  array catygroies from database 
 let theBooks;
 let eidtFalg=false;
 let idOfrowHolderForEdit; // variable to hold the id of catygory to edit;
+let canBeDeleted=true;
+let imgName;
+
+
+function canBeDeletedOrNot(id)
+{
+     theBooks.forEach(element => {
+       
+        if(element.authorId==id)
+        {
+            canBeDeleted=false;
+            
+        }
+        if(element.categoryId==id)
+        {
+            canBeDeleted=false;
+           
+        }
+
+    });
+
+}
+
 
 function addingNewCatygoryfunction()
 {
@@ -104,9 +133,19 @@ function editingAuthor(event)
     let rowAuthor=event.target.parentNode.parentNode.rowIndex
     let theActionToTake=event.target.id;
     if(theActionToTake=="AuthordeleteBtn")
-    {   //please write if condition here mohamed 
+  {  
+      canBeDeletedOrNot(theAuthors[rowAuthor-1]._id)
+      if(canBeDeleted==true)
+        {
         AuthorTable.deleteRow(rowAuthor)
         senddeletedDataToserver(theAuthors[rowAuthor-1]._id,'author')
+       
+        }
+        else
+        {
+        $("#modalCannotBeDeleted").modal("show")
+        }
+        canBeDeleted=true;
     }
     else if(theActionToTake=="AuthorEditBtn")
     {
@@ -115,10 +154,10 @@ function editingAuthor(event)
     let inputAuthorBoxName=document.getElementById("inputAuthorBoxName")
     let inputAuthorLastNameBox=document.getElementById("inputAuthorLastNameBox")
     let inputAuthorDateOfBirthBox=document.getElementById("inputAuthorDateOfBirthBox")
-    let inputImageForTheAuthorBox=document.getElementById("inputImageForTheAuthorBox") //waiting for karim
     inputAuthorBoxName.value=AuthorTable.rows[rowAuthor].cells[0].innerHTML;
     inputAuthorLastNameBox.value=AuthorTable.rows[rowAuthor].cells[1].innerHTML;
-    inputAuthorDateOfBirthBox.value=AuthorTable.rows[rowAuthor].cells[2].innerHTML;
+    inputAuthorBioGraphyBox.value=AuthorTable.rows[rowAuthor].cells[2].innerHTML
+    inputAuthorDateOfBirthBox.value=AuthorTable.rows[rowAuthor].cells[3].innerHTML;
      $('#modalAuthor').modal('show');
     }
     displayingAllTable()
@@ -131,9 +170,19 @@ function editingCatygory(event)
    let theActionToTake=event.target.id;
    if(theActionToTake=="catygorydeleteBtn")
    {
-       CatygroiesTable.deleteRow(rowCat);
+    canBeDeletedOrNot(theCartygroies[rowCat-1]._id)
+    if(canBeDeleted==true)
+    {
+     CatygroiesTable.deleteRow(rowCat);
        senddeletedDataToserver(theCartygroies[rowCat-1]._id,'category')
-   }
+       canBeDeleted=true;
+    }
+    else
+    {
+        $("#modalCannotBeDeleted").modal("show")
+    }
+    
+  }
     else if(theActionToTake=="catygoryEditBtn")
    {
     eidtFalg=true;
@@ -194,11 +243,13 @@ function showingTheAuthorDiv()
 }
 function displayTheCatygroiesIndropDown()
 {
+    inputCatTypForTheBookBox.innerHTML="";
     let id=0;
      theCartygroies.forEach(element => {
      inputCatTypForTheBookBox.innerHTML+=' <option value='+id+'>'+element.name+'</option>'
           id++;
     });
+    theAuthors.innerHTML="";
     id=0;
     theAuthors.forEach(element => {
         inputAuthorTypForTheBookBox.innerHTML+=' <option value='+id+'>'+element.first_name+'</option>'
@@ -210,8 +261,8 @@ function addingNewAuthor()
     let inputAuthorBoxName=document.getElementById("inputAuthorBoxName").value
     let inputAuthorLastNameBox=document.getElementById("inputAuthorLastNameBox").value
     let inputAuthorDateOfBirthBox=document.getElementById("inputAuthorDateOfBirthBox").value
-    let inputImageForTheAuthorBox=document.getElementById("inputImageForTheAuthorBox").value
-    if((inputAuthorBoxName==""||inputAuthorLastNameBox==""||inputAuthorDateOfBirthBox=="")&&eidtFalg==false)
+    let authorBio=inputAuthorBioGraphyBox.value;
+    if((inputAuthorBoxName==""||inputAuthorLastNameBox==""||inputAuthorDateOfBirthBox==""||inputImageForTheAuthorBox.value=="")&&eidtFalg==false)
     {
       pleaseEnterDataAuthor.style.display="block";
     }
@@ -222,11 +273,13 @@ function addingNewAuthor()
            first_name:inputAuthorBoxName,
            last_name:inputAuthorLastNameBox,
            dateOfBirth:inputAuthorDateOfBirthBox,
-           photoName:"waitingforupload", // waiting for karim
+           photoName:imgName, 
+           biography:authorBio,
           }
           if(eidtFalg==false)
         {   
            sendDataToserver(newAuthor,'author')
+           console.log(newAuthor)
         }
         else 
         {
@@ -240,15 +293,15 @@ function addingNewAuthor()
 function displayAuthorData()
 {
     
-    AuthorTable.innerHTML='<thead><tr><th scope="col">First name</th><th scope="col">Last name</th><th scope="col">Date of birth</th><th scope="col">Image</th><th  class="text-center" scope="col">Actions</th></tr></thead><tbody>';
+    AuthorTable.innerHTML='<thead><tr><th scope="col">First name</th><th scope="col">Last name</th><th scope="col">Biography</th><th scope="col">Date of birth</th><th scope="col">Image</th><th  class="text-center" scope="col">Actions</th></tr></thead><tbody>';
     theAuthors.forEach(element => {
-        AuthorTable.innerHTML+='<tr><td>'+element.first_name+'</td><td>'+element.last_name+'</td><td>'+element.dateOfBirth+'</td><td>'+element.photoName+'</td><td  class="text-center"><input type="button" id="AuthorEditBtn" class="btn btn-info btn-xs " value="Edit"><input type="button" class="btn btn-danger btn-xs marginToTheLeft"  id="AuthordeleteBtn" value="Delete"></td></tr>'
+        AuthorTable.innerHTML+='<tr><td>'+element.first_name+'</td><td>'+element.last_name+'</td><td>'+element.biography+'</td><td>'+element.dateOfBirth+'</td><td><img width="50px;" hieght="50px;" src="img/'+element.photoName+'"></td><td  class="text-center"><input type="button" id="AuthorEditBtn" class="btn btn-info btn-xs " value="Edit"><input type="button" class="btn btn-danger btn-xs marginToTheLeft"  id="AuthordeleteBtn" value="Delete"></td></tr>'
     });
     AuthorTable.innerHTML+='</tbody>'
 }
 function displayTheAuthorsTable()
 {
-   fetch('http://localhost:3000/author',
+   fetch('http://localhost:3000/author/Author',
     {
        method:"GET",
        headers: {Accept: 'application/json'},
@@ -266,27 +319,30 @@ function addingNewBookFn()
    name=inputBookBoxName.value;
    catOfbook=inputCatTypForTheBookBox.value;
    authorOfthebook=inputAuthorTypForTheBookBox.value;
-  if(name==""||catOfbook==""||authorOfthebook=="")
+   bookDescrib=inputBookBoxDescription.value;
+  if(name==""||catOfbook==""||authorOfthebook==""||inputImageForTheBookBox.value=="")
   { pleaseEnterDataBook.style.display="block";}
   else
   {     let newBook={
         name:name,
         categoryId:theCartygroies[catOfbook]._id,
         authorId:theAuthors[authorOfthebook]._id,
-        photoName:"waiting for karim",
+        photoName:imgName,
+        description:bookDescrib,
+
    }
     if(eidtFalg==false)
     {
-    sendDataToserver(newBook,'book');
+       sendDataToserver(newBook,'book');
     }else if(eidtFalg==true)
     {
         sendEditeddataToserver(idOfrowHolderForEdit,'book',newBook)
     }
     displayingAllTable();
     $('#modalBook').modal('hide');
-   }
-   
+ }
   
+   
 }
 
 
@@ -308,6 +364,7 @@ function editingOrDeletingBooks(event)
         idOfrowHolderForEdit=theBooks[currentRow-1]._id;
         displayTheCatygroiesIndropDown()  
         inputBookBoxName.value=bookTable.rows[currentRow].cells[0].innerHTML;
+        inputBookBoxDescription.value=bookTable.rows[currentRow].cells[4].innerHTML;
         $("#modalBook").modal("show")
       }
 }
@@ -315,7 +372,7 @@ function editingOrDeletingBooks(event)
 
 async function displayThebookTable()
 {
-    await fetch('http://localhost:3000/book',
+    await fetch('http://localhost:3000/book/books',
     {
        method:"GET",
        headers: {Accept: 'application/json'},
@@ -324,16 +381,16 @@ async function displayThebookTable()
       return res.json();
    }).then ( data => {
     theBooks=data;
-    
+    console.log("hyy")
     displayBookInTable();})
 }
 function displayBookInTable()
 {
    
-   bookTable.innerHTML='<thead><tr><th scope="col">Name</th><th scope="col">Catygorie</th><th scope="col">Author</th><th scope="col">Image</th><th  class="text-center" scope="col">Actions</th></tr></thead><tbody>'
+   bookTable.innerHTML='<thead><tr><th scope="col">Name</th><th scope="col">Catygorie</th><th scope="col">Author</th><th scope="col">Image</th><th scope="col">Description</th><th  class="text-center" scope="col">Actions</th></tr></thead><tbody>'
 
 theBooks.forEach(element => {
-bookTable.innerHTML+='<tr><td>'+element.name+'</td><td>'+gettingtheNameOfTheCat(element.categoryId)+'</td><td>'+gettingtheNameOfTheAuthor(element.authorId)+'</td><td>'+element.photoName+'</td><td  class="text-center"><input type="button" id="bookEditBtn" class="btn btn-info btn-xs " value="Edit"><input type="button" class="btn btn-danger btn-xs marginToTheLeft"  id="bookdeleteBtn" value="Delete"></td></tr>'
+bookTable.innerHTML+='<tr><td>'+element.name+'</td><td>'+gettingtheNameOfTheCat(element.categoryId)+'</td><td>'+gettingtheNameOfTheAuthor(element.authorId)+'</td><td><img width="50px;" hieght="50px;" src="img/'+element.photoName+'"></td><td>'+element.description+'</td><td  class="text-center"><input type="button" id="bookEditBtn" class="btn btn-info btn-xs " value="Edit"><input type="button" class="btn btn-danger btn-xs marginToTheLeft"  id="bookdeleteBtn" value="Delete"></td></tr>'
 });
 }
 function gettingtheNameOfTheCat(id)
@@ -384,21 +441,74 @@ $("#modalCat").on("hidden.bs.modal", function () {
   });
   $("#modalBook").on("hidden.bs.modal", function () {
     eidtFalg=false;
-   
+    inputBookBoxDescription.value="";
+    inputBookBoxName.value="";
+    inputImageForTheBookBox.value="";
+    
   });
 $("#modalAuthor").on("hidden.bs.modal", function () {
     eidtFalg=false;
     let inputAuthorBoxName=document.getElementById("inputAuthorBoxName")
     let inputAuthorLastNameBox=document.getElementById("inputAuthorLastNameBox")
     let inputAuthorDateOfBirthBox=document.getElementById("inputAuthorDateOfBirthBox")
-    let inputImageForTheAuthorBox=document.getElementById("inputImageForTheAuthorBox")
+    inputAuthorBioGraphyBox.value=""
     inputAuthorBoxName.value="";
     inputAuthorLastNameBox.value="";
     inputAuthorDateOfBirthBox.value="";
-    inputImageForTheAuthorBox="";
+    inputImageForTheAuthorBox.value="";
     pleaseEnterDataAuthor.style.display="none";
 });
 
   
 
- 
+uploadBtnAuthorImage.addEventListener("click" , (event)=>
+{
+  
+  let imgPathArr = inputImageForTheAuthorBox.value.split("\\");
+   imgName = imgPathArr[imgPathArr.length - 1];
+
+  let formData = new FormData();
+  formData.append('photo', inputImageForTheAuthorBox.files[0] ,imgName);
+    
+    fetch("http://localhost:3000/user/image",
+    {
+        method: 'POST',
+        body:formData
+    }).then (res =>res.text() )
+    .then( response => {
+        if(response == "done")
+        {
+            uploadBtnAuthorImage.style.backgroundColor="green";
+            uploadBtnAuthorImage.style.color="black";
+        } else {
+            uploadBtnAuthorImage.style.backgroundColor="red";
+            uploadBtnAuthorImage.style.color="black";
+        }
+    })
+})
+
+uploadBtnBookImage.addEventListener("click" , (event)=>
+{
+  
+  let imgPathArr = inputImageForTheBookBox.value.split("\\");
+   imgName = imgPathArr[imgPathArr.length - 1];
+
+  let formData = new FormData();
+  formData.append('photo', inputImageForTheBookBox.files[0] ,imgName);
+    
+    fetch("http://localhost:3000/user/image",
+    {
+        method: 'POST',
+        body:formData
+    }).then (res =>res.text() )
+    .then( response => {
+        if(response == "done")
+        {
+            uploadBtnBookImage.style.backgroundColor="green";
+            uploadBtnBookImage.style.color="black";
+        } else {
+            uploadBtnBookImage.style.backgroundColor="red";
+            uploadBtnBookImage.style.color="black";
+        }
+    })
+})
